Add files via upload
Start of authentication app to be integrated with ranking system.    We may or may not start with a register system for the new user, if not we can remove once it functional.  There is a tutorial that walks you through how to complete this (express-cc chris youtube)     note: .env file is on gitignore due to safety so the file will need to be recreated.  Node modules have been removed also.  Starts up with nodemon and webpack.